Refactor time
