+++ src/time/time.pyi
